backend/requirements.txt
Django==5.2.5 
djangorestframework==3.16.1 
djangorestframework-simplejwt==5.5.1 
django-cors-headers==4.7.0 
gunicorn==23.0.0 
psycopg2-binary==2.9.10 
python-whois==0.9.5 
requests==2.32.4 
dj-database-url==2.1.0 
whitenoise==6.12.0 
